requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==21.4b0
chardet==4.0.0
click==7.1.2
discord.py==1.7.1
idna==3.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.17.0
regex==2021.4.4
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
